Annotators should take a simple io.Reader
Switched around Buffer content storage from [][]StyledRune to a simple
byte array, storing the positions of newlines, highlights, and
annotations as byte offsets into that array. This should cut down on
memory usage and clean up the Annotator interface by eliminating the
need for domain-specific objects like BufferReader and PositionedRune.

In the process, a Stream object was decomposed from Buffer. A Stream is
essentially a byte array that synchronizes concurrent reading/writing,
blocking on read until either enough data is available or the stream is
closed.