More conventional dotfile loading
1. Drop the '.' in '.ratrc'. Redundant since the parent directory is
hidden. Fixes #16.

2. Instead of hardcoding the location of rat's config directory, use
$XDG_CONFIG_HOME, following the convention on Unix-like systems.
  https://en.wikipedia.org/wiki/Freedesktop.org
  https://specifications.freedesktop.org/basedir-spec/latest

In the process I've created a new file for dotfile issues, which may be
useful for supporting #17 (loading multiple config files).

Open question: should we update the default `flags.cmd`? I've minimally
updated it for now, but it's still useless to anyone overriding
$XDG_CONFIG_HOME. I'm new to Go, and I don't know how to pass
non-literal defaults to flags. I also wonder if the default should be
something else.