Support reading from STDIN
If `cmd` is not provided, read from STDIN.

This allows you to pipe things into rat. For example,

```
$ git log --oneline | rat -m 'git.log'
$ ls -al | rat -m 'files'
```

This is kind of convenient, but there is one downside. Because rat
doesn't know what command generated the content, you will not be able to
reload pagers that read from STDIN.

Github issue #31