Implement configurable auto-reload behavior
Use the `reload` keyword inside of a mode definition to specify a
command that will be run for each pager in that mode. Every line output
from that process to STDOUT will trigger a reload of the pager.

Debounce with a constant number for now. This number may end up being
configurable some day...